fix: move ssrf exception to App\Exceptions and mark guard readonly to satisfy arch tests
