+++ app/Exceptions/SsrfGuardException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Services\Favicon;
+namespace App\Exceptions;
 
 use RuntimeException;
