+++ app/Services/Favicon/Drivers/HighQualityDriver.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace App\Services\Favicon\Drivers;
 
+use App\Exceptions\SsrfGuardException;
 use App\Services\Favicon\SsrfGuard;
-use App\Services\Favicon\SsrfGuardException;
 use AshAllenDesign\FaviconFetcher\Collections\FaviconCollection;
 use AshAllenDesign\FaviconFetcher\Concerns\HasDefaultFunctionality;
 use AshAllenDesign\FaviconFetcher\Concerns\MakesHttpRequests;
